初步实现大颗粒算法

+++ core/stream/algorithm/model/instance.py
@@ -2,7 +2,7 @@
 class Instance:
     '''模型结果实例'''
     def __init__(self, label_id: int, label_name: str, score: float,
-                 left: float, top: float, right: float, bottom: float):
+                 left: int, top: int, right: int, bottom: int):
         self.label_id = label_id
         self.label_name = label_name
         self.score = score
